requirements.txt
Flask==3.1.2
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
injector==0.22.0
langchain_openai==1.1.0
python-dotenv==1.2.1
SQLAlchemy==2.0.44
WTForms==3.2.1
